Moving setOptGroup into essence20
